requirements.txt
pytest
click
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
WTForms
nltk
scikit-learn
sklearn-crfsuite